Changed the backend and added a trigger function
A lot of errors were encountered when changing the schema, relationships, and then adding a lambda function.
The AWS documentation was very poorly written and verbose in some cases, which didn't help in finding the solution to my problem.
My goal was to server side sort a list of users and their groups they are in, so when I query, I will get a sorted list.
The problem lies with relationships. I was not able to create a secondary sort key, based on the 3rd table that is linked with the join table.
I had to create a trigger so when ChatGroup is updated, the join table updatedAt field, between User and ChatGroup, would also be updated.
I can then query via user and create a key for createdAt in the join table.