Finish adding attachments
I had a lot of problems with gql and then updating the messages when a new
attachment was created. I opted to just subscribe to any created attachments
and then update the first value of the message. My problem was that both the
message and attachments are created on the same time/function. I may have to
change how it works in the future to create a more separate approach.